refactor: migrate to feature-based package structure
Replace layer-based packages (models/, repo/, services/, controllers/, DTO/)
with feature packages (admin/, doctor/, patient/, appointment/, prescription/,
shared/). Each domain now owns its model, repository, service, and controller.
Cross-cutting services (TokenService, ValidationService, Service) and DTOs
(AuthDTO) move to shared/. Add shared/event/ with AppointmentBookedEvent and
AppointmentCancelledEvent records as Kafka message-contract placeholders.
