+++ app/src/main/java/com/project/back_end/admin/Admin.java
@@ -1,4 +1,4 @@
-package com.project.back_end.models;
+package com.project.back_end.admin;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
