+++ app/src/main/java/com/project/back_end/appointment/AppointmentRepository.java
@@ -1,4 +1,4 @@
-package com.project.back_end.repo;
+package com.project.back_end.appointment;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -9,8 +9,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.project.back_end.models.Appointment;
 
 @Repository
 public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
