+++ app/src/main/java/com/project/back_end/appointment/AppointmentService.java
@@ -1,4 +1,4 @@
-package com.project.back_end.services;
+package com.project.back_end.appointment;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -10,14 +10,12 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.project.back_end.DTO.AppointmentDTO;
 import com.project.back_end.enums.ServiceResult;
-import com.project.back_end.models.Appointment;
-import com.project.back_end.models.Doctor;
-import com.project.back_end.models.Patient;
-import com.project.back_end.repo.AppointmentRepository;
-import com.project.back_end.repo.DoctorRepository;
-import com.project.back_end.repo.PatientRepository;
+import com.project.back_end.doctor.Doctor;
+import com.project.back_end.doctor.DoctorRepository;
+import com.project.back_end.patient.Patient;
+import com.project.back_end.patient.PatientRepository;
+import com.project.back_end.shared.TokenService;
 
 @Service
 public class AppointmentService {
