+++ app/src/main/java/com/project/back_end/doctor/DoctorRepository.java
@@ -1,12 +1,10 @@
-package com.project.back_end.repo;
+package com.project.back_end.doctor;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-
-import com.project.back_end.models.Doctor;
 
 @Repository
 public interface DoctorRepository extends JpaRepository<Doctor, Long> {
