+++ app/src/main/java/com/project/back_end/mvc/DashboardController.java
@@ -1,6 +1,6 @@
 package com.project.back_end.mvc;
 
-import com.project.back_end.services.Service;
+import com.project.back_end.shared.Service;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
