+++ app/src/main/java/com/project/back_end/patient/PatientRepository.java
@@ -1,6 +1,4 @@
-package com.project.back_end.repo;
-
-import com.project.back_end.models.Patient;
+package com.project.back_end.patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
