+++ app/src/main/java/com/project/back_end/prescription/PrescriptionRepository.java
@@ -1,11 +1,9 @@
-package com.project.back_end.repo;
+package com.project.back_end.prescription;
 
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-
-import com.project.back_end.models.Prescription;
 
 @Repository
 public interface PrescriptionRepository extends MongoRepository<Prescription, String>  {
