+++ app/src/main/java/com/project/back_end/shared/AuthDTO.java
@@ -1,4 +1,4 @@
-package com.project.back_end.DTO;
+package com.project.back_end.shared;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
