+++ app/src/main/java/com/project/back_end/shared/event/AppointmentBookedEvent.java
@@ -1,0 +1,11 @@
+package com.project.back_end.shared.event;
+
+import java.time.LocalDateTime;
+
+// Kafka message payload — publish after a successful appointment booking.
+public record AppointmentBookedEvent(
+    long appointmentId,
+    long patientId,
+    long doctorId,
+    LocalDateTime appointmentTime
+) {}
