+++ app/src/main/java/com/project/back_end/shared/event/AppointmentCancelledEvent.java
@@ -1,0 +1,7 @@
+package com.project.back_end.shared.event;
+
+// Kafka message payload — publish after a successful appointment cancellation.
+public record AppointmentCancelledEvent(
+    long appointmentId,
+    long patientId
+) {}
