+++ app/src/test/java/com/project/back_end/controllers/AppointmentControllerTest.java
@@ -1,15 +1,16 @@
 package com.project.back_end.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.project.back_end.DTO.AppointmentDTO;
+import com.project.back_end.appointment.Appointment;
+import com.project.back_end.appointment.AppointmentController;
+import com.project.back_end.appointment.AppointmentDTO;
+import com.project.back_end.appointment.AppointmentService;
 import com.project.back_end.enums.ServiceResult;
 import com.project.back_end.exceptions.InvalidTokenException;
-import com.project.back_end.models.Appointment;
-import com.project.back_end.models.Doctor;
-import com.project.back_end.models.Patient;
-import com.project.back_end.services.AppointmentService;
-import com.project.back_end.services.Service;
-import com.project.back_end.services.ValidationService;
+import com.project.back_end.doctor.Doctor;
+import com.project.back_end.patient.Patient;
+import com.project.back_end.shared.Service;
+import com.project.back_end.shared.ValidationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
